+++ perl/src/main/java/io/perl/impl/HdrExtendedLatencyRecorder.java
@@ -15,7 +15,7 @@ import io.perl.LatencyRecord;
 import io.perl.LatencyRecordWindow;
 import io.perl.LatencyRecorder;
 import io.perl.ReportLatencies;
-import io.perl.system.PerlPrinter;
+import io.perl.PerlPrinter;
 import io.time.Time;
 import org.HdrHistogram.Histogram;
 import org.HdrHistogram.HistogramIterationValue;
